moved fullscreen-method to store and moved all the saveTo-methods to the context-menu component
